pyproject.toml
[tool.poetry]
name = "djangoproject6"
version = "0.1.0"
description = ""
authors = ["Павел <king-he@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.5"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
